// src/components/types/BtnStackProps.tsx
export interface BtnStackProps {
  border?: string
  height?: string
  minHeight?: string
  maxHeight?: string
  minWidth?: string
  maxWidth?: string
  width?: string
  display?: string
  direction?: string | number | any
  align?: string
  justifyContent?: string
  gridColumns?: string
  gridRows?: string
  placeContent?: string
  whiteSpace?: string | number | any
  wrap?: string | number | any
  overflow?: string
  shadow?: string
  children: any
}